[Refactor] refactor modal and children
